requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pandas==2.2.2
numpy==1.26.4
pydantic==2.9.1
python-dotenv==1.0.1
google-genai==1.66.0
